requirements.txt
Django==1.7.5
PyJWT==0.4.1
Unipath==1.1
django-extensions==1.5.0
django-social-auth==0.7.28
httplib2==0.9
oauth2==1.5.211
oauthlib==0.7.2
python-instagram==1.3.0
python-openid==2.2.5
python-social-auth==0.2.2
requests==2.5.3
requests-oauthlib==0.4.2
simplejson==3.6.5
six==1.9.0
wsgiref==0.1.2
